// src/app/@core/api/model/adrDto.ts
/**
 * EPR api
 * The list of EPR project endpoints
 *
 * OpenAPI spec version: 1.0
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */
import { ProjetDto } from '././projetDto';
import { ZoneDto } from '././zoneDto';


export interface AdrDto { 
    id: number;
    status: AdrDto.StatusEnum;
    date: string;
    commentaire: string;
    conforme: boolean;
    rapporteur: string;
    adr: string;
    local: string;
    intitule: string;
    responsable: string;
    projet: ProjetDto;
    zone: ZoneDto;
}
export namespace AdrDto {
    export type StatusEnum = 'in progress' | 'finished';
    export const StatusEnum = {
        InProgress: 'in progress' as StatusEnum,
        Finished: 'finished' as StatusEnum
    };
}